requirements.txt
python-telegram-bot==22.2
aiohttp==3.12.14